requirements.txt
flask
pandas
scikit-learn
numpy
mysql-connector-python